update registration and validation user

+++ src/UI/Action/DeleteTrickAction.php
@@ -2,11 +2,9 @@
 
 namespace App\UI\Action;
 
-use App\Domain\Models\Trick;
 use App\Domain\Repository\TrickRepository;
 use App\UI\Responder\DeleteTrickResponder;
 use Symfony\Component\HttpFoundation\Request;
-use Doctrine\Common\Persistence\ObjectManager;
 use Symfony\Component\Routing\Annotation\Route;
 use App\UI\Action\Interfaces\DeleteTrickActionInterface;
 
@@ -28,11 +26,15 @@ class DeleteTrickAction implements DeleteTrickActionInterface
      * 
      * @Route("/supprimerTrick/{id}", name="deltrick", methods={"GET"})
      */
-    public function __invoke(ObjectManager $manager, Request $request, DeleteTrickResponder $responder)
+    public function __invoke(Request $request, DeleteTrickResponder $responder)
     {
 
-        $this->trickRepository->delete($request->get('id'));
+        if($request->get('id')){
 
-        return $responder();
+            $this->trickRepository->delete($request->get('id'));
+
+            return $responder();
+        }
+
     }
 }
